refactor(Line): extract exploring sections
- 과도한 책임을 줄이기 위해 조회 관련 기능을 모두 LineSectionExplorer로 이관
- LineSectionExplorer는 외부에 노출할 도메인은 아니기 때문에 모든 메서드를 기본 접근자로 제한
