+++ src/main/java/nextstep/subway/line/domain/LineSectionExplorer.java
@@ -4,6 +4,7 @@ import nextstep.subway.line.domain.exceptions.ExploreSectionException;
 import nextstep.subway.station.domain.Station;
 
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
@@ -12,13 +13,33 @@ import java.util.stream.Collectors;
 public class LineSectionExplorer {
     private final List<Section> sections;
 
-    public LineSectionExplorer(final List<Section> sections) {
+    LineSectionExplorer(final List<Section> sections) {
         this.sections = new ArrayList<>(sections);
     }
 
-    public Station findUpStation() {
+    Station findUpStation() {
         Section theFirstSection = findFirstSection();
         return theFirstSection.getUpStation();
+    }
+
+    List<Station> getStations() {
+        if (sections.isEmpty()) {
+            return Arrays.asList();
+        }
+
+        LineSectionExplorer lineSectionExplorer = new LineSectionExplorer(sections);
+
+        List<Station> stations = new ArrayList<>();
+        stations.add(lineSectionExplorer.findUpStation());
+
+        Section currentSection = lineSectionExplorer.findFirstSection();
+
+        while (currentSection != null) {
+            stations.add(currentSection.getDownStation());
+            currentSection = this.findNextSection(currentSection);
+        }
+
+        return stations;
     }
 
     Section findFirstSection() {
@@ -38,5 +59,12 @@ public class LineSectionExplorer {
                 .filter(it -> it.getValue() == 1L)
                 .map(Map.Entry::getKey)
                 .collect(Collectors.toList());
+    }
+
+    Section findNextSection(Section currentSection) {
+        return this.sections.stream()
+                .filter(it -> currentSection.getDownStation().equals(it.getUpStation()))
+                .findFirst()
+                .orElse(null);
     }
 }
